--- a/src/rust/engine/engine_pyo3/src/externs/interface.rs
+++ b/src/rust/engine/engine_pyo3/src/externs/interface.rs
@@ -4,11 +4,22 @@
 use pyo3::exceptions::PyException;
 use pyo3::prelude::*;
 
+mod nailgun;
 mod testutil;
 
 #[pymodule]
-fn native_engine_pyo3(_py: Python, m: &PyModule) -> PyResult<()> {
+fn native_engine_pyo3(py: Python, m: &PyModule) -> PyResult<()> {
   m.add_class::<PyExecutor>()?;
+
+  m.add(
+    "PantsdConnectionException",
+    py.get_type::<self::nailgun::PantsdConnectionException>(),
+  )?;
+  m.add(
+    "PantsdClientException",
+    py.get_type::<self::nailgun::PantsdClientException>(),
+  )?;
+  m.add_class::<self::nailgun::PyNailgunClient>()?;
 
   m.add_class::<self::testutil::PyStubCAS>()?;
   m.add_class::<self::testutil::PyStubCASBuilder>()?;