--- a/src/rust/engine/engine_pyo3/src/externs/interface.rs
+++ b/src/rust/engine/engine_pyo3/src/externs/interface.rs
@@ -9,20 +9,10 @@ mod testutil;
 
 #[pymodule]
 fn native_engine_pyo3(py: Python, m: &PyModule) -> PyResult<()> {
+  self::nailgun::register(py, m)?;
+  self::testutil::register(m)?;
+
   m.add_class::<PyExecutor>()?;
-
-  m.add(
-    "PantsdConnectionException",
-    py.get_type::<self::nailgun::PantsdConnectionException>(),
-  )?;
-  m.add(
-    "PantsdClientException",
-    py.get_type::<self::nailgun::PantsdClientException>(),
-  )?;
-  m.add_class::<self::nailgun::PyNailgunClient>()?;
-
-  m.add_class::<self::testutil::PyStubCAS>()?;
-  m.add_class::<self::testutil::PyStubCASBuilder>()?;
 
   Ok(())
 }