We no longer use the `with_executor` mechanism from https://github.com/pantsbuild/pants/issues/11722, as we can directly store an `Executor` on the `PyNailgunClient` cheaply. This allows us to use `Python.allow_threads()` for better parallelism.

It's awkward that our Python code now has `PyExecutor` from Rust-Cpython and PyO3, and that those aren't compatible. But this is less offensive than the risk of one big change rather than an incremental migration.

We can't yet migrate `PyNailgunSession` because `PyCancellationLatch` is used by `PySession`.

--

This also improves our file organization. We use a new pattern of having a `register()` function in each `interface/` file to register that file's functions and classes on the single native extension module, which makes `interface.rs` less bloated.