@Eric-Arellano 2021-06-09
Can you think of a better benchmark? Iiuc, this executes the PyNailgun code every run. I wanted to avoid too much noise from things like launching subprocesses. I also disabled BuildSense, remote caching, and telemetry.

Before:

```
❯ hyperfine --warmup 1 --runs 100 './pants list ::'
Benchmark #1: ./pants list ::
  Time (mean ± σ):     404.9 ms ±  11.4 ms    [User: 272.4 ms, System: 53.4 ms]
  Range (min … max):   396.9 ms … 463.6 ms    100 runs
```

After:

```
❯ hyperfine --warmup 1 --runs 100 './pants list ::'
Benchmark #1: ./pants list ::
  Time (mean ± σ):     400.4 ms ±  17.4 ms    [User: 267.9 ms, System: 52.8 ms]
  Range (min … max):   393.7 ms … 518.3 ms    100 runs
```

If this is a sensible benchmark, I think this shows an overall, albeit small improvement? The variance makes that harder to reason about. But probably small enough of an impact for this PR to be safe to merge?