@stuhood 2021-06-09
There really aren't any hot loops in here, so I'd trust a benchmark that showed no change.